requirements.txt
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43
pandas==2.3.2
python-dotenv==1.1.1
requests==2.32.5